Update patch set 5
Patch Set 5:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > For some reason, I cannot push an update to gerrit right now. I will pick this up later today and hopefully, I can push then.
> 
> Probably this: https://www.eclipse.org/lists/cross-project-issues-dev/msg17797.html

Yes probably, it works now anyway.

I figured out why it failed and I've modified the script to try to do pre-processing of the source file(s) in a step before doing the actual build and the pre-processed content is then hashed. This way, all the include files are hashed so if anything changes that have an effect on the binary, it will get caught there. I also added hashing of the command line used for building the library/executable.

Patch-set: 5