Update patch set 12
Patch Set 12: Code-Review+2

> Patch Set 12: Code-Review+2
> 
> (4 comments)
> 
> This looks good to me. It does not interfere with normal operation.
> 
> Without knowing much about flatpak I do worry a little about the detection method (just env variable). I couldn't find docs about FLATPAK_SANDBOX_DIR - but I did find plenty of third-party references to it.

This seems to be the widespread way for detecting it.

> 
> Can someone run Eclipse CDT entirely, including target code in the flatpak so that they wouldn't want to flatpak-spawn?

In theory this is possible if eclipse flatpak uses flatpak runtime that contains the whole toolchain (so called SDK flatpak runtimes) but this makes cdt install enormous as these sdk runtimes contain everything needed to build the binary runtime incl. headers. So there is possibility for this to happen but at least for now I haven't seen anyone ship their own SDK runtime and the existing ones are used to build e.g. Gnome/KDE/Freedesktop runtimes which are huge and just waiting space for people not particularly developing Gnome or KDE apps. It's a good idea to provide some way for user to opt out of flatpak spawn but it's not as important as being able to use it so it can wait for next release IMHO.

> 
> Can someone have another reason to have the env variable set? 

I can't think of such reason.


> 
> If I try with FLATPAK_SANDBOX_DIR set doing build and run give errors visibly in the console (e.g. make all 
> Cannot run program "flatpak-spawn": Unknown reason) - but doing a Debug just leaves an unremovable launch in the Debug view with the error being in the error log.

flatpak-spawn is available only inside the sandbox so that's normal. We should still guard it though to reduce the chance for people to see such errors.

> 
> I think all the above is minor, so please submit it and we can resolve the minor issues I have raised between M3 and release.

Patch-set: 12
Reviewer: Gerrit User 726 <726@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=+2