Update patch set 2
Patch Set 2:

(2 comments)

> Patch Set 1:
> > 
> > I would change implementation to extract the following steps:
> I'm unsure, which implementation you are referring to here. Couuld you explain?

I'm referring to "CMakeErrorParser extends OutputStream" - it does too much for output stream, and I would extract the dedicated types for each step enumerated below - we will have smaller, more focused and better testable classes with clear contract.

> 
> > 1) read buffer
> > 2) use patterns to create "marker info" - it can be a map with attributes
> > 3) determine resource to add marker
> > 4) create marker
> > after that the code will be more simple and more testable.

To be more specific, I would change `processMessage` to return an info sufficient to fulfill IMarker (let's say CMakeMarkerInfo or just Map) and then I would add a Consumer<CMakeMarkerInfo> to be a final field of CMakeErrorParser.
From this point we are not bound to IResource anymore and we can easily test any input for CMakeErrorParser with plain JUnit, because we can substitute Consumer<CMakeMarkerInfo> with any IResource-free implementation. The same approach we can use for "determine resource to add marker" step and so on. On every step we will have simple, cohesive and testable code instead of one big OutputStream successor that implicitly requeres IWorkspace to work with CMake command line ooutput.

Of cause we can postpone this rework for the next release and go ahead with what we have in PS2.

Patch-set: 2
Reviewer: Gerrit User 23745 <23745@97ee7c02-f12f-4043-b43e-dea463d88b31>