Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> (2 comments)
> 
> > Patch Set 1:
> > > 
> > > I would change implementation to extract the following steps:
> > I'm unsure, which implementation you are referring to here. Couuld you explain?
> 
> I'm referring to "CMakeErrorParser extends OutputStream" - it does too much for output stream, and I would extract the dedicated types for each step enumerated below - we will have smaller, more focused and better testable classes with clear contract.
> 
> > 
> > > 1) read buffer
> > > 2) use patterns to create "marker info" - it can be a map with attributes
> > > 3) determine resource to add marker
> > > 4) create marker
> > > after that the code will be more simple and more testable.
> 
> To be more specific, I would change `processMessage` to return an info sufficient to fulfill IMarker (let's say CMakeMarkerInfo or just Map) and then I would add a Consumer<CMakeMarkerInfo> to be a final field of CMakeErrorParser.
> From this point we are not bound to IResource anymore and we can easily test any input for CMakeErrorParser with plain JUnit, because we can substitute Consumer<CMakeMarkerInfo> with any IResource-free implementation. The same approach we can use for "determine resource to add marker" step and so on. On every step we will have simple, cohesive and testable code instead of one big OutputStream successor that implicitly requeres IWorkspace to work with CMake command line ooutput.
> 
> Of cause we can postpone this rework for the next release and go ahead with what we have in PS2.

Thank you for your insights. I now agree this should be re-worked to be more maintenance/testing friendly.
The current code was taken from cmake4eclipse and is in production for years, so I think it has been field-tested.

Let Jonah decide whether this should be in the next release or not. If rework is going to be postponed, there should be a bug report assigned to me. Once bug 
566385 is resolved w/o the rework, I will start reworking this.

To hook in the error parser, a follow-up, non-trival gerrit is required in any case because cmake is run again in the 'make' phase if the user modified one of the cmakelists.txt since the last build. 
But that output is impossible to parse (stdout and stderr lines come in intermixed). Cmake4eclipse solved that by tracking changes to the cmakelists.txt files and forcing cmake to run in advance then with the error parser attached. We need file-modification detection here; CBuildConfiguration.elementChanged() looks promising, but you never know...

Patch-set: 2