Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (3 comments)
> 
> Well done on tackling some code that is very old in CDT. Some of this stuff dates back to 2002 - such as the "PersistentPTY.getInputStream() has non-API return type PTYInputStream" 
> 
> Unfortunately this code is used by many other projects and I think it is just too late (after M3) to make such breaking changes. 
> 
> Annoyingly this is a problem in part because consumers, such as org.eclipse.remote, were non-specific in their imports regarding versions.
> 
> I am open to being convinced, especially if there is a benefit in this change beyond just the cleanup.
> 
> In addition, as this has API breakages, it needs to be documented in doc/org.eclipse.cdt.doc.isv/guide/deprecated_API_removals.html

I did not create a ticket for this since I would like to have some feedback if this was the right way forward or of it should be left as is first.
If there is such a high risk of breaking stuff, I guess none of the changes are candidate for inclusion. Is there anything in the gerrit that you think we should merge or should I abandon the entire changeset?

Patch-set: 1