Update patch set 1
Patch Set 1:

This is a slightly larger changeset, but I think it's in the trivial ballpark anyway.

I did not do this by hand. I used a small script to extract the strings and place them in the .properties file and then update the MANIFEST.MF file with the proper tag.

Patch-set: 1
Reviewer: Gerrit User 27769 <27769@97ee7c02-f12f-4043-b43e-dea463d88b31>