Update patch set 1
Patch Set 1:

Currently, there are a few different ways to handle translations in CDT code base.
One is using the attribute "Bundle-Localization" in MANIFEST.MF with a file containing the attribute value.properties in the root of the bundle.
Another is to not define the attribute in MANIFEST.MF and place the translation in OSGI-INF/l10n/bundle.properties.

Maybe it's time to converge on one of these strategies and stick to it?

Also, I notice that the default encoding is configured to ISO-8859-1 in Eclipse, but from Java9, the new default encoding is UTF-8. I tried to locate a setting in Eclipse that controls this, but I was unable. Anyone know where it is or if there even is such a setting?

Patch-set: 1