Update patch set 3
Patch Set 3:

As far as I understand test 40 it is creating a project based on a zip file and compares that to a set of hardcoded (in the test) data.
I can't find \${ in the zip nor in the hardcoded data.
The difference that leads to a failed test is in the include directives missing in the created project. I found the include directives in the .cproject but I'm not used to reading these files. but nothing pointed me to \§{

Patch-set: 3