Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> I can reproduce by importing that zip file (org.eclipse.cdt/build/org.eclipse.cdt.managedbuilder.core.tests/resources/test40Projects/test_40/test_40.zip) into CDT and building with and without your patch. With the patch "${workspace_loc:/test_40/dir1/dir2/dir3}" is being discarded from the list of strings, without it it is expanded to the workspace_loc.
> 
> I couldn't simply reproduce it in a new project with just "${workspace_loc:/test_40/dir1/dir2/dir3}" in includes, so it may have to do with the full list of includes in that variable, or something else.

I think I found the issue. I escapes 2 characters in the regex for the varname but escaping is not needed resulting in the fact the varname can not hold a \.
I don't quite see how that gets us here, but the original code did not backslash ^and { in the regex.

Patch-set: 3