Update patch set 2
Patch Set 2:

(1 comment)

Time is incredibly short for LSP4E to have a new build for 2020-09. This is because the LSP4E project does not do milestones as a matter of course. There was some discussion of it earlier and I have re-raised it.

Patch-set: 2
Reviewer: Gerrit User 27769 <27769@97ee7c02-f12f-4043-b43e-dea463d88b31>