Update patch set 2
Patch Set 2:

In continuing this discussion - the shortest path is probably a global switch to get it in CDT 10 with the refinements in this change and the other one (https://git.eclipse.org/r/c/cdt/org.eclipse.cdt/+/168415) being rolled out in due course.

I think it is totally acceptable for the experimental release to have a "requires restart" for the preference. This will allow interested parties to experiment with the new code without breaking other people.

I hope that in a short period of time that the LSP version can be on by default and the code just do the right thing.

Patch-set: 2
Reviewer: Gerrit User 27769 <27769@97ee7c02-f12f-4043-b43e-dea463d88b31>