Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> Currently, there are a few different ways to handle translations in CDT code base.
> One is using the attribute "Bundle-Localization" in MANIFEST.MF with a file containing the attribute value.properties in the root of the bundle.
> Another is to not define the attribute in MANIFEST.MF and place the translation in OSGI-INF/l10n/bundle.properties.

I think this is old vs new way - OSGI-INF is the new way. I am not sure what the advantages of the new way are. I have added Alexander to CC because he has been championing that in the past.

> 
> Maybe it's time to converge on one of these strategies and stick to it?

Sure - keep in mind that moving translation files should have a update to all existing translations too and that happens outside of CDT in the babel project, it is something I have never really engaged in, just vaguely familiar that it exists: https://www.eclipse.org/babel/

> 
> Also, I notice that the default encoding is configured to ISO-8859-1 in Eclipse, but from Java9, the new default encoding is UTF-8. I tried to locate a setting in Eclipse that controls this, but I was unable. Anyone know where it is or if there even is such a setting?

See Bug 566511

Patch-set: 2
CC: Gerrit User 23745 <23745@97ee7c02-f12f-4043-b43e-dea463d88b31>