Update patch set 1
Patch Set 1:

Hi Jonah.  There was a bug reported by Vaclav about opening the profiling dialog which turned out to be a reference to CMainTab in our plugin.xml.  I was fooling around with a new Core Build Makefile project and it was failing because the template Makefile did not support build modes that weren't "run" or "debug" and it also had a bug in that an $(error) statement was being flagged as a recipe by Makefile as it started with a tab.

I have posted a fix here that removes the $(error) message and adds some build flags to support gcov and gprof when a user profiles.  There is still an issue that if you build run or debug and right-click on the binary and select code coverage, it will build correctly but will still think the binary isn't ready for code coverage (separate build folders).  Right-clicking from the profiling build directory will work after that or using the launchbar to profile works fine.  This gcov issue is something that can be left for the next release.  I can remove the gcov and gprof flags if you want and force a user to manually add the flags (i.e. compile flags will just be -g).

Patch-set: 1
Reviewer: Gerrit User 27769 <27769@97ee7c02-f12f-4043-b43e-dea463d88b31>