Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> This looks a little weird to me. i.e. if a user does "make" at the command line, rather than error it will create a profiling exe. 
> 
> It is fine for now though and can be revisited in the future one the profile mode can be passed down.

If you like, I can change it to look for BUILD_MODE = linuxtools (the mode we set) and add the error statement back for nothing specified or I can make the default simply -g and force the user to make the changes manually when needed.  Both are easy to do.

Patch-set: 1