Bug 567169 - Detection for VS 2017 and VS2019 toolchain
With this change, support for VS 2008, 2010 and 2012 is dropped
and support for 2017 and 2019 is added.

The new detection uses vswhere.exe as we cannot rely on registry keys
anymore. We also retrieve correct values for INCLUDE, LIB, and PATH
by executing vcvars.bat, which is less brittle and hard-coded than
before. Doing this also partially addresses bug 357980.

We also now store knowledge about multiple VS installation (one for
each major version) and related MSVC toolchain. This is one step closer
to letting users choose which version to use and store this per
configuration in the project.

Change-Id: I3b3579d8c6742ef232626b5e5294bb6f8634a326
Signed-off-by: Marc-Andre Laperle <malaperle@gmail.com>