Update patch set 2
Patch Set 2: Code-Review-1

Michael, thank you for the contribution!

CDT's parser code tries to (roughly) follow the structure of the C++ specification (available for easy browsing at http://eel.is/c++draft/). This helps ensure that CDT's behaviour matches that of the spec (modulo bugs / unimplemented features), and makes it easier to track down issues where CDT's behaviour diverges from the spec and determine how to fix them.

So, when fixing a bug where CDT's behaviour diverges from the spec, the first step is generally to understand what part(s) of the spec lead to the intended outcome in the first place.

For the code example in the bug, the issue is the call site in testLValue2(), where CDT believes the two overloads of demo() are ambiguous, but according to the spec the second one is the better overload.

I believe the clause that's relevant to this outcome is http://eel.is/c++draft/over.match.best#general-2.5: the second overload of demo() is "more specialized than" the first overload for the purposes of this clause.

More specifically, following the link in that clause to http://eel.is/c++draft/temp.func.order, and from there to http://eel.is/c++draft/temp.deduct.partial, I believe it is http://eel.is/c++draft/temp.deduct.partial#9.1 that ends up making the second overload more specialized.

In CDT code, the implementation of http://eel.is/c++draft/temp.func.order is in CPPTemplates.orderFunctionTemplates(). So, I believe the fix should go inside the implementation of orderFunctionTemplates() (or one of its helpers, as appropriate), and the added code should match the logic of http://eel.is/c++draft/temp.deduct.partial#9.1.

Patch-set: 2
Reviewer: Gerrit User 973 <973@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=-1