Update patch set 3
Patch Set 3:

> Patch Set 2:
> 
> @Jonah do you have any objections to switch to 4.18-I and to Orbit 2020-12?

I-builds used to be non-tested so I've had quite a few bad builds in the past (but was happy to participate in helping Platform testing as part of my previous job!). Not sure we want to necessarily expose all CDT contributors to that. But it's not like we can't change to another build easily if there's a problem and there is less activity in the platform...so maybe it's OK.

Patch-set: 3
CC: Gerrit User 598 <598@97ee7c02-f12f-4043-b43e-dea463d88b31>