Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> It looks like we changed pixel-based magic numbers to character-based magic numbers.
> How it will work if we will decide to change text? What about translated strings?

The magic numbers are dimensions for text entry boxes only, not for labels. In each case we have a good idea of the maximum number of chars for these fields. For example, an address offset will be a 32-bit or 64-bit hex number, so we know how many digit chars are required.

Patch-set: 1