Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> > Patch Set 5:
> > 
> > > Patch Set 5:
> > > 
> > > Why does code-cleanlyness want to change
> > > Copyright (c) 2015, 2020
> > > into
> > > Copyright (c) 2020, 2020 ?
> 
> > In the future we may need a different way to handle this. Suggestions welcome.
> 
> Suggestion: If it is a move-operation, skip that check.

Thanks - what I need is a suggestion of how to detect a move-operation as git does not record moves explicitly. --follow is not reliable as git finds similarity with other files and most feature.xml's can be followed all the way back to the very first one.

Anyway, the workaround now is fine, and perhaps one day so many things will be worked around that the removing the check will be easier. The check only needed adding because there was such a mess and inconsistencies in this area in the past.

Patch-set: 5