Update patch set 2
Patch Set 2:

(1 comment)

> It also seems to turn into a self-confirming one (by age right), which upsets me a bit.

No it does not - I just don't want to see gerrits filled with non-relevant changes. If you think something in CDT should be changed you have been empowered as a committer and explicitly to raise bugs and improve them. Age does not do it - the objection is that your issues are off-topic. When many off-topic issues are regularly raised in a gerrit the usefulness of the review diminishes.

I have provided a long answer to your other query about logs in this gerrit. For the sake of having an easier to review gerrit going forward I am abandoning this one so that the issues relevant to this gerrit can be identified.

If there are any issues remaining, please either raise bugs if off topic, or add new comment to the new gerrit I will post soon.

Patch-set: 2