Bug 568375: Ensure compatible versions of javax.xml.bind and com.sun.xml.bind are used
It is important that the version of javax.xml.bind is the same version
as com.sun.xml.bind.

Note at the time of writing, there is only one version of each of these
in orbit. See the bug report as the com.sun.xml.bind provided by orbit
needs an update too.

Change-Id: I38c21c5c2045ec18d917bdded9ac1afb11ee86c1