Update patch set 2
Patch Set 2:

This looks good. I haven't tried it yet, but I was wondering if this can be done with enablements of the Eclipse tracing system so that it is easier in the future to remember how to turn it on? I think that would mean using a method to initialize/set tracing options, or have the JNI code read the debug setting - not sure what would be easier.

Patch-set: 2
Reviewer: Gerrit User 27769 <27769@97ee7c02-f12f-4043-b43e-dea463d88b31>