Update patch set 3
Patch Set 3:

> Patch Set 2:
> 
> This looks good. I haven't tried it yet, but I was wondering if this can be done with enablements of the Eclipse tracing system so that it is easier in the future to remember how to turn it on? I think that would mean using a method to initialize/set tracing options, or have the JNI code read the debug setting - not sure what would be easier.

That was my intention, but I stranded on the problem that I either need to have an init method in the native code that is exposed to Java world or that function call in the native code will need to have the JNIEnv object passed as it's needed to access Java objects and their methods/fields.
If you know any global method that can be used to get the JNIEnv instance from anywhere in the native code, it would certainly be an improvement.
The main problem I see with the init method is the risk of a race.

I've also tried to play with the environment variables and having them set in Java world before loading the DLL, but it does not look like there is a way to modify the environment for the running process, just for the children to the java process...

Patch-set: 3