Update patch set 5
Patch Set 5:

> Patch Set 3:
> 
> I am not sure how much detail I should review this in. If there is something you would like me to pay attention to, please let me know, other LGTM (except missing bug number)

Nothing special, just want some kind of confidence that there is not obvious errors.
A few more of the error states are captured and logged and the allocated resources are now returned rather than leaked, so I guess there could be something I've overlooked. For the use-cases that I've been testing, it does appear to work as intended.

There are a few casts between jchar*, wchar_t* and char* a little allover. I'm not sure they are safe, but they were there earlier too.
With the changes that I've introduced, I think it should be possible to do some kind of unittest approach, but I have not looked into what the usual way to do that is with JNI. Maybe it's not worth the extra work to do that and simply just run the IDE and see if it works...? As there are no unitests for the native code, I suggest that it will be addressed in a different ticket if it's decided to go that way.

Patch-set: 5