Update patch set 2
Patch Set 2:

(4 comments)

> Patch Set 2:
> 
> Ping @alexander: Waiting for your valuable input...

@Martin you know my comments in advance 😊 I would prefer to see smaller cohesive immutable classes that can be easily tested outside the heavy framework lifecycle. 
In the added code I can see the logic of argument parsing and analysis that can be extracted (to be tested by unit tests) relatively easy.

I do not consider this as "must have before the merge", so it is up to you when to improve the implementation.

Patch-set: 2
Reviewer: Gerrit User 23745 <23745@97ee7c02-f12f-4043-b43e-dea463d88b31>