Update patch set 4
Patch Set 4:

(2 comments)

@Jonah, please review this and if possible, git it a spin and see what you think.

I havn't tried this on Windows as I have to wait for a complete build to be ready to run on the borrowed system, but I see no reason why it should fail on Windows now with this minor change. Change set 4 works fine on Linux anyway.

Patch-set: 4