Update patch set 1
Patch Set 1:

This change set would improve performance if https://git.eclipse.org/r/c/cdt/org.eclipse.cdt/+/171780 is accepted. The problem with this one is that it does not build due to that the added methods in WindowsRegistry does not exist in DuseSimrelRepo. Anyway around this obstacle or should the gerrit be split into two parts, one for the new functions and one for the use of them? I guess a simrel would be needed in between too...?

Patch-set: 1
Reviewer: Gerrit User 27769 <27769@97ee7c02-f12f-4043-b43e-dea463d88b31>