Update patch set 4
Patch Set 4:

(2 comments)

This is an interesting approach, it is actually somewhat the opposite of what I would expect the solution to be however. Based on the bug description, I would have guessed that what the user actually types should take preference. So if the user changes "make -j2" to "make -j3" and changes to the Behavior tab, then in that tab the setting would display as 3 parallel. TBH I would go so far as to say if a user unchecks "Use default build command" that the "Build settings" part of Behavior tab be disabled.

Something like adding to Builder.setArguments() the removal of any -j option, and if the found -j option does not match the builder setting, update the setting to match so it can be reconstructed. But then the UI needs to reflect that change too.

However after looking at this code there probably would be a mountain to climb to achieve the above. 

There are a few cases that this code does not handle. For example, if user types "-j 3" then this code removes the -j but leaves the 3. Or if a user types "--jobs=3" then this new code is not triggered at all.

Patch-set: 4
Reviewer: Gerrit User 27769 <27769@97ee7c02-f12f-4043-b43e-dea463d88b31>