Bug 568616 - Support for __is_same and __is_same_as (built-in equivalent to std::is_same)
IType.isSameType was doing pretty much already what was needed. Except
there was an issue when using it in combination with __underlying_type,
where it would end up being a CPPUnaryTypeTransformation being compared
even though this class should only be for the case of being used on a
dependent type.

Change-Id: Ifd2908e726c098fb07c9420b29e2cb26014419bf
Signed-off-by: Marc-Andre Laperle <malaperle@gmail.com>