Update patch set 4
Patch Set 4:

Thanks for the contribution. Even though this does not seem to be the solution at the moment the high level discussion seems to have helped us arrive at a solution.

PS I was trying to only address issues with your change, the Builder code has lots of other problems and I apologize if anything I pointed to was not actually related to your change.

Patch-set: 4