Update patch set 2
Patch Set 2:

Have you thought of making a subclass of InstaciationContext (e.g. TemplateArgumentInstantiationContext) to avoid the if-statement in TypeInstantiationRequest and to avoid to enrich the InstnantiationContext with additional logic (isForDeduction)

In my opinion it would be better and clearer to encapsulate this whole stuff in the News class TemplateArgumentInstantiationContext.

(Sorry if I did not see all spelling mistakes. My German Android phone makes mistakes which I didn't type.)

Patch-set: 2
CC: Gerrit User 301437 <301437@97ee7c02-f12f-4043-b43e-dea463d88b31>