Update patch set 3
Patch Set 3:

I'm not 100% sure about this change. I've used this build command field in the case to replace 'make' with a drop-in replacement, like 'gmake'. Perhaps one could also use field to specify an absolute path to 'make'. Then in those case you'd want to preserve the other flags as configured in the UI. As it is, this change would likely break existing configs and probably warrants an entry in N&N as projects would need to be updated to account for this.

Patch-set: 3
CC: Gerrit User 598 <598@97ee7c02-f12f-4043-b43e-dea463d88b31>