Update patch set 3
Patch Set 3:

Hmm. I think I might have found another issue now.

The following files needs to have +x permission if we go with enforcement:
   core/org.eclipse.cdt.core.tests/resources/testlib/x86/so.g/libtestlib_g.so
   debug/org.eclipse.cdt.debug.application/scripts/cdtdebug.sh
   debug/org.eclipse.cdt.debug.application/scripts/install.sh
   llvm/org.eclipse.cdt.managedbuilder.llvm.ui/scripts/find_path.bat
   llvm/org.eclipse.cdt.managedbuilder.llvm.ui/scripts/stdlib_path.sh
   releng/scripts/cleanup.py

Looking at the list, I was starting to wonder how the stdlib_path.sh and find_path.bat scripts are delivered. If they are delivered using a .jar file, the permissions will be lost at the user end of the transmission. Furthermore, the files are referenced from java code, but they are not included in the build.properties, so I guess they are not part of the delivery?!
There is actually a comment in llvm/org.eclipse.cdt.managedbuilder.llvm.ui/src/org/eclipse/cdt/managedbuilder/llvm/util/FindStdLibPath.java that more or less coveres this problem with permissions in jar files.
Also worth noting here is that it's likely that the commands wont work as quite a few Linux distributions no longer use "locate".

I guess some of the above is also valid for the cdtdebug.sh and install.sh.

cleanup.py should probably not have execute permission as it looks like it's supposed to be executed inside some context.

Please advise what I should do 😊

Patch-set: 3