Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> Hmm. I think I might have found another issue now.

:-L I wasn't that attached to the +x enforced, so allow is probably fine...

> 
> The following files needs to have +x permission if we go with enforcement:
>    core/org.eclipse.cdt.core.tests/resources/testlib/x86/so.g/libtestlib_g.so
>    debug/org.eclipse.cdt.debug.application/scripts/cdtdebug.sh
>    debug/org.eclipse.cdt.debug.application/scripts/install.sh
>    llvm/org.eclipse.cdt.managedbuilder.llvm.ui/scripts/find_path.bat
>    llvm/org.eclipse.cdt.managedbuilder.llvm.ui/scripts/stdlib_path.sh
>    releng/scripts/cleanup.py

... however I don't see a problem with any of these having +x in the git repo.

> 
> Looking at the list, I was starting to wonder how the stdlib_path.sh and find_path.bat scripts are delivered. If they are delivered using a .jar file, the permissions will be lost at the user end of the transmission. Furthermore, the files are referenced from java code, but they are not included in the build.properties, so I guess they are not part of the delivery?!
> There is actually a comment in llvm/org.eclipse.cdt.managedbuilder.llvm.ui/src/org/eclipse/cdt/managedbuilder/llvm/util/FindStdLibPath.java that more or less coveres this problem with permissions in jar files.
> Also worth noting here is that it's likely that the commands wont work as quite a few Linux distributions no longer use "locate".

For the above comment I would create a new bug (especially about the locate issue) and CC Marc-Andre.

> 
> I guess some of the above is also valid for the cdtdebug.sh and install.sh.

These scripts are used by https://wiki.eclipse.org/CDT/StandaloneDebugger#Try_out_Full_Eclipse_Installation which say to run install.sh with sh so it does not need to be executable, so it doesn't matter if the executable bit is lost. Also install.sh makes cdtdebug.sh executable as part of the install. So I don't think it matters what state it is in the git repo.

> 
> cleanup.py should probably not have execute permission as it looks like it's supposed to be executed inside some context.

It is - it is probably less relevant now. It was quite important when I had to rebase a number of gerrits and client forks onto a reformatted CDT: https://kichwacoders.com/2018/11/24/cdt-has-been-reformatted/

> 
> Please advise what I should do 😊

I think you should make it all executable, unless there is still some unforeseen problem or something I did not understand. Also, I think the original problem you were solving (cygwin vs egit) is the key thing, so whatever is best for that should have a big influence.

Has that helped?

Patch-set: 3